recycling removed since it is not the primary source of memory consumption according to profiling (hprof).
bugs fixed in Trie and SurfaceDictionary.

external k-value buffer (kernel function results) added for svm predicting -- one primary source of memory consumption (~20%).

stored procedures updated to eliminate transferring strings -- another primary source of memory consumption (~30%).

libsvm code is incorporated; minor changes to implement external k-value buffer.

+++ ecologylab/semantics/concept/database/DatabaseFacade.java
@@ -193,7 +193,7 @@ public class DatabaseFacade extends Debug
 		}
 		return preparedStatements.get(sql);
 	}
-
+	
 	/**
 	 * query for keyphraseness from the database.
 	 * 
@@ -277,10 +277,10 @@ public class DatabaseFacade extends Debug
 		try
 		{
 			CallableStatement cst = conn.prepareCall("{ ? = call query_inlink_count(?) }");
-			cst.registerOutParameter(1, Types.BIGINT);
+			cst.registerOutParameter(1, Types.INTEGER);
 			cst.setString(2, toConcept);
 			cst.execute();
-			inlinkCount = (int) cst.getLong(1);
+			inlinkCount = cst.getInt(1);
 			cst.close();
 		}
 		catch (SQLException e)
